plot updates
