+++ regression/scripts/B737/mission_B737.py
@@ -17,9 +17,14 @@
 # ----------------------------------------------------------------------
 
 import SUAVE
-from SUAVE.Core import Units
-from SUAVE.Visualization.Performance.Mission_Plots import *
-from SUAVE.Visualization.Geometry import * 
+from SUAVE.Core import Units 
+from SUAVE.Visualization.Performance.Aerodynamics.Vehicle import *  
+from SUAVE.Visualization.Performance.Mission              import *  
+from SUAVE.Visualization.Performance.Energy.Common        import *  
+from SUAVE.Visualization.Performance.Energy.Battery       import *   
+from SUAVE.Visualization.Performance.Energy.Fuel          import *  
+from SUAVE.Visualization.Performance.Noise                import * 
+from SUAVE.Visualization.Geometry.Three_Dimensional       import * 
 import matplotlib.pyplot as plt  
 import numpy as np 
 
@@ -189,28 +194,28 @@ def base_analysis(vehicle):
 #   Plot Mission
 # ----------------------------------------------------------------------
 
-def plot_mission(results,line_style='bo-'):
+def plot_mission(results):
     
     # Plot Flight Conditions 
-    plot_flight_conditions(results, line_style)
+    plot_flight_conditions(results)
     
     # Plot Aerodynamic Forces 
-    plot_aerodynamic_forces(results, line_style)
+    plot_aerodynamic_forces(results)
     
     # Plot Aerodynamic Coefficients 
-    plot_aerodynamic_coefficients(results, line_style)
+    plot_aerodynamic_coefficients(results)
     
     # Plot Static Stability Coefficients 
-    plot_stability_coefficients(results, line_style)    
+    plot_stability_coefficients(results)    
     
     # Drag Components
-    plot_drag_components(results, line_style)
+    plot_drag_components(results)
     
     # Plot Altitude, sfc, vehicle weight 
-    plot_altitude_sfc_weight(results, line_style)
+    plot_altitude_sfc_weight(results)
     
     # Plot Velocities 
-    plot_aircraft_velocities(results, line_style)  
+    plot_aircraft_velocities(results)  
 
     return
