+++ regression/scripts/Embraer_E190_constThr/mission_Embraer_E190_constThr.py
@@ -15,7 +15,12 @@
 
 import SUAVE
 from SUAVE.Core import Units
-from SUAVE.Visualization.Performance.Mission_Plots import *  
+from SUAVE.Visualization.Performance.Aerodynamics.Vehicle import *  
+from SUAVE.Visualization.Performance.Mission              import *  
+from SUAVE.Visualization.Performance.Energy.Common        import *  
+from SUAVE.Visualization.Performance.Energy.Battery       import *   
+from SUAVE.Visualization.Performance.Energy.Fuel          import *  
+from SUAVE.Visualization.Performance.Noise                import *  
 import matplotlib.pyplot as plt  
 import numpy as np  
 
@@ -393,17 +398,17 @@ def mission_setup(analyses):
 #   Plot Mission
 # ----------------------------------------------------------------------
 
-def plot_mission(results,line_style='bo-'):
+def plot_mission(results):
 
-    plot_altitude_sfc_weight(results, line_style) 
+    plot_altitude_sfc_weight(results) 
     
-    plot_flight_conditions(results, line_style) 
+    plot_flight_conditions(results) 
     
-    plot_aerodynamic_coefficients(results, line_style)  
+    plot_aerodynamic_coefficients(results)  
     
-    plot_aircraft_velocities(results, line_style)
+    plot_aircraft_velocities(results)
     
-    plot_drag_components(results, line_style)
+    plot_drag_components(results)
 
     return
