+++ regression/scripts/airfoil_analysis/airfoil_panel_method_test.py
@@ -13,7 +13,7 @@ from SUAVE.Methods.Geometry.Two_Dimensional.Cross_Section.Airfoil.compute_naca_4
      import  compute_naca_4series
 from SUAVE.Methods.Geometry.Two_Dimensional.Cross_Section.Airfoil.import_airfoil_geometry\
      import import_airfoil_geometry
-from SUAVE.Visualization.Performance.Airfoil_Aerodynamics import * 
+from SUAVE.Visualization.Performance.Aerodynamics.Airfoil import * 
 import os 
 import numpy as np
