+++ regression/scripts/nonuniform_propeller_inflow/nonuniform_propeller_inflow.py
@@ -6,7 +6,7 @@
 import SUAVE
 from SUAVE.Core import Units, Data
 from SUAVE.Methods.Propulsion import propeller_design
-from SUAVE.Visualization.Performance.Propeller_Plots import *
+from SUAVE.Visualization.Performance.Aerodynamics.Rotor import *
 from SUAVE.Methods.Aerodynamics.Common.Fidelity_Zero.Lift.compute_wing_wake import compute_wing_wake
 from SUAVE.Methods.Aerodynamics.Common.Fidelity_Zero.Lift.compute_propeller_nonuniform_freestream import compute_propeller_nonuniform_freestream
