+++ regression/scripts/segments/segment_test.py
@@ -636,7 +636,6 @@ def missions_setup(base_mission):
 
     # done!
     return missions  
-
-    return  
+ 
 if __name__ == '__main__': 
     main()
