+++ regression/scripts/slipstream/propeller_interactions.py
@@ -14,7 +14,7 @@ from SUAVE.Analyses.Propulsion.Rotor_Wake_Fidelity_One import Rotor_Wake_Fidelit
 from SUAVE.Methods.Propulsion.Rotor_Wake.Fidelity_One.compute_wake_induced_velocity import compute_wake_induced_velocity
 from SUAVE.Methods.Aerodynamics.Common.Fidelity_Zero.Lift.compute_propeller_nonuniform_freestream import compute_propeller_nonuniform_freestream
 from SUAVE.Methods.Aerodynamics.Common.Fidelity_Zero.Lift.generate_propeller_grid import generate_propeller_grid
-from SUAVE.Visualization.Performance.Propeller_Plots import *
+from SUAVE.Visualization.Performance.Aerodynamics.Rotor import *
 
 import numpy as np
 import pylab as plt
