+++ trunk/SUAVE/Visualization/Performance/Aerodynamics/Airfoil/__init__.py
@@ -1,6 +1,7 @@
-## @defgroup Visualization-Performance-Airfoil_Aerodynamics Performance 
-# @ingroup Visualization 
-
+## @defgroup Visualization-Performance-Aerodynamics-Airfoil Airfoil
+# Plotting Functions for Airfoils
+# @ingroup Visualization-Performance-Aerodynamics
+  
 from .plot_airfoil_boundary_layer_properties    import plot_airfoil_boundary_layer_properties
 from .plot_airfoil_polar_files                  import plot_airfoil_polar_files
 from .plot_airfoil_surface_forces               import plot_airfoil_surface_forces
