+++ trunk/SUAVE/Visualization/Performance/Aerodynamics/Rotor/__init__.py
@@ -1,0 +1,7 @@
+## @defgroup Visualization-Performance-Aerodynamics-Rotor Rotor
+# Plotting Functions for Rotors
+# @ingroup Visualization-Performance-Aerodynamics
+ 
+from .plot_rotor_disc_inflow                    import plot_rotor_disc_inflow 
+from .plot_rotor_disc_performance               import plot_rotor_disc_performance
+from .plot_rotor_performance                    import plot_rotor_performance
