+++ trunk/SUAVE/Visualization/Performance/Aerodynamics/Vehicle/__init__.py
@@ -1,0 +1,9 @@
+## @defgroup Visualization-Performance-Aerodynamics-Vehicle Vehicle
+# Plotting Functions for the Vehicle
+# @ingroup Visualization-Performance-Vehicle
+
+from .plot_aerodynamic_coefficients             import plot_aerodynamic_coefficients
+from .plot_aerodynamic_forces                   import plot_aerodynamic_forces
+from .plot_drag_components                      import plot_drag_components
+from .plot_lift_distribution                    import plot_lift_distribution
+from .plot_stability_coefficients               import plot_stability_coefficients
