+++ trunk/SUAVE/Visualization/Performance/Aerodynamics/Vehicle/plot_aerodynamic_coefficients.py
@@ -1,5 +1,5 @@
 ## @ingroup Visualization-Performance-Aerodynamics
-# plot_stability_coefficients.py
+# plot_aerodynamic_coefficients.py
 # 
 # Created:    Nov 2022, E. Botero
 # Modified:   
@@ -18,22 +18,24 @@ import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 
 # ---------------------------------------------------------------------- 
-#   Stability Coefficients
+#   Aerodynamic Coefficients
 # ---------------------------------------------------------------------- 
 
 ## @ingroup Visualization-Performance-Aerodynamics
-def plot_stability_coefficients(results,
-                                save_figure=False,
-                                save_filename="Stability_Coefficents",
-                                file_type=".png",
-                                width = 1600, height =800,
-                                *args, **kwargs):
-    """This plots the static stability characteristics of an aircraft
+def plot_aerodynamic_coefficients(results,
+                                  save_figure=False,
+                                  save_filename="Aerodynamic_Coefficents",
+                                  file_type=".png",
+                                  width = 1400, height = 665,
+                                  *args, **kwargs):
+    """This plots the aerodynamic coefficients
     
     Assumptions:
     None
     
     Source:
+
+    Deprecated SUAVE Mission Plots Functions
 
     Created:    Mar 2020, M. Clarke
     Modified:   Apr 2020, M. Clarke
@@ -41,51 +43,48 @@ def plot_stability_coefficients(results,
                 Apr 2021, M. Clarke
                 Dec 2021, S. Claridge
     
-    
     Inputs:
-    results.segments.conditions.stability.
-       static
-           CM
-           Cm_alpha
-           static_margin
-       aerodynamics.
-           angle_of_attack
+    results.segments.condtions.aerodynamics.
+        lift_coefficient
+        drag_coefficient
+        angle_of_attack
+        
     Outputs:
-    
     Plots
+    
     Properties Used:
     N/A
     """
-    
-    plot_cols = ['CM',
-                 'CM_a',
-                 'SM',
+
+    plot_cols = ['CL',
+                 'CD',
                  'AoA',
+                 'L_D',
                  'Segment']
 
     df = pd.DataFrame(columns=plot_cols)
-    
 
+    # Get the segment-by-segment results
     for segment in results.segments.values():
-        time     = segment.conditions.frames.inertial.time[:,0] / Units.min
-        cm       = segment.conditions.stability.static.CM[:,0]
-        cm_alpha = segment.conditions.stability.static.Cm_alpha[:,0]
-        SM       = segment.conditions.stability.static.static_margin[:,0]
-        aoa      = segment.conditions.aerodynamics.angle_of_attack[:,0] / Units.deg
+        time = segment.conditions.frames.inertial.time[:,0] / Units.min
+        cl   = segment.conditions.aerodynamics.lift_coefficient[:,0,None]
+        cd   = segment.conditions.aerodynamics.drag_coefficient[:,0,None]
+        aoa  = segment.conditions.aerodynamics.angle_of_attack[:,0] / Units.deg
+        l_d  = cl/cd    
         
         segment_frame = pd.DataFrame(
-        np.column_stack((cm,
-                         cm_alpha,
-                         SM,
-                         aoa)),
+        np.column_stack((cl,
+                         cd,
+                         aoa,
+                         l_d)),
         columns = plot_cols[:-1], index=time
         )
 
         segment_frame['Segment'] = [segment.tag for i in range(len(time))]
     
         # Append to collecting data frame
-        df = df.append(segment_frame)             
-
+        df = df.append(segment_frame)        
+        
     # Set plot parameters
     fig = make_subplots(rows=2, cols=2)
 
@@ -97,42 +96,42 @@ def plot_stability_coefficients(results,
             x=data.index,
             y=data['AoA'],
             name=seg_name),
-            row=1, col=1)    
+            row=1, col=1)
         
         fig.add_trace(go.Scatter(
             x=data.index,
-            y=data['CM'],
+            y=data['CL'],
             name=seg_name,
             showlegend=False),
-            row=1, col=2)        
-
+            row=1, col=2)
+        
         fig.add_trace(go.Scatter(
             x=data.index,
-            y=data['CM_a'],
+            y=data['CD'],
+            name=seg_name,
+            showlegend=False),
+            row=2, col=2)
+        
+        fig.add_trace(go.Scatter(
+            x=data.index,
+            y=data['L_D'],
             name=seg_name,
             showlegend=False),
             row=2, col=1)        
-
-        fig.add_trace(go.Scatter(
-            x=data.index,
-            y=data['SM'],
-            name=seg_name,
-            showlegend=False),
-            row=2, col=2)        
-
+    
     fig.update_yaxes(title_text='AoA (deg)', row=1, col=1)
-    fig.update_yaxes(title_text='$C_M$', row=1, col=2)
-    fig.update_yaxes(title_text='$C_M\alpha$', row=2, col=1)
-    fig.update_yaxes(title_text='Static Margin (%)', row=2, col=2)
+    fig.update_yaxes(title_text='CL', row=1, col=2)
+    fig.update_yaxes(title_text='CD', row=2, col=2)
+    fig.update_yaxes(title_text='L/D', row=2, col=1)
     
     fig.update_xaxes(title_text='Time (min)', row=2, col=1)
-    fig.update_xaxes(title_text='Time (min)', row=2, col=2)        
+    fig.update_xaxes(title_text='Time (min)', row=2, col=2)    
 
     # Set overall figure layout style and legend title
     fig.update_layout(
         width=width, height=height,
         legend_title_text='Segment',
-        title_text = 'Stability Coefficients'
+        title_text = 'Aerodynamic Coefficients'
     )
 
     fig = plot_style(fig)
