+++ trunk/SUAVE/Visualization/Performance/Aerodynamics/Vehicle/plot_aerodynamic_forces.py
@@ -1,5 +1,5 @@
 ## @ingroup Visualization-Performance-Aerodynamics
-# plot_aerodynamic_coefficients.py
+# plot_aerodynamic_forces.py
 # 
 # Created:    Nov 2022, E. Botero
 # Modified:   
@@ -18,23 +18,23 @@ import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 
 # ---------------------------------------------------------------------- 
-#   Aerodynamic Coefficients
+#   Aerodynamic Forces
 # ---------------------------------------------------------------------- 
 
 ## @ingroup Visualization-Performance-Aerodynamics
-def plot_aerodynamic_coefficients(results,
-                                  save_figure=False,
-                                  save_filename="Aerodynamic_Coefficents",
-                                  file_type=".png",
-                                  width = 1600, height = 665,
-                                  *args, **kwargs):
-    """This plots the aerodynamic coefficients
+def plot_aerodynamic_forces(results,
+                            save_figure=False,
+                            save_filename="Aerodynamic_Forces",
+                            file_type=".png",
+                            width = 1400, height = 665,
+                            *args, **kwargs):
+    """This plots the aerodynamic forces
     
     Assumptions:
     None
     
     Source:
-
+    
     Deprecated SUAVE Mission Plots Functions
 
     Created:    Mar 2020, M. Clarke
@@ -44,94 +44,95 @@ def plot_aerodynamic_coefficients(results,
                 Dec 2021, S. Claridge
     
     Inputs:
-    results.segments.condtions.aerodynamics.
-        lift_coefficient
-        drag_coefficient
-        angle_of_attack
-        
+    results.segments.condtions.frames
+         body.thrust_force_vector
+         wind.lift_force_vector
+         wind.drag_force_vector
+         
     Outputs:
     Plots
     
     Properties Used:
     N/A
     """
-
-    plot_cols = ['CL',
-                 'CD',
-                 'AoA',
-                 'L_D',
+    
+    # Create empty dataframe to be populated by the segment data
+    plot_cols = ['Thrust',
+                 'Lift',
+                 'Drag',
+                 'eta',
                  'Segment']
 
-    df = pd.DataFrame(columns=plot_cols)
+    df = pd.DataFrame(columns=plot_cols)    
 
-    # Get the segment-by-segment results
     for segment in results.segments.values():
-        time = segment.conditions.frames.inertial.time[:,0] / Units.min
-        cl   = segment.conditions.aerodynamics.lift_coefficient[:,0,None]
-        cd   = segment.conditions.aerodynamics.drag_coefficient[:,0,None]
-        aoa  = segment.conditions.aerodynamics.angle_of_attack[:,0] / Units.deg
-        l_d  = cl/cd    
+        time   = segment.conditions.frames.inertial.time[:,0] / Units.min
+        Thrust = segment.conditions.frames.body.thrust_force_vector[:,0]
+        Lift   = -segment.conditions.frames.wind.lift_force_vector[:,2]
+        Drag   = -segment.conditions.frames.wind.drag_force_vector[:,0]
+        eta    = segment.conditions.propulsion.throttle[:,0]
         
+        # Assemble data into temporary holding data frame
         segment_frame = pd.DataFrame(
-        np.column_stack((cl,
-                         cd,
-                         aoa,
-                         l_d)),
-        columns = plot_cols[:-1], index=time
+            np.column_stack((Thrust,
+                             Lift,
+                             Drag,
+                             eta)),
+            columns = plot_cols[:-1], index=time
         )
-
         segment_frame['Segment'] = [segment.tag for i in range(len(time))]
-    
+
         # Append to collecting data frame
         df = df.append(segment_frame)        
-        
+
+
     # Set plot parameters
     fig = make_subplots(rows=2, cols=2)
-
+    
     # Add traces to the figure for each value by segment
     for seg, data in df.groupby("Segment", sort=False):
-        seg_name = ' '.join(seg.split("_")).capitalize()  
+        seg_name = ' '.join(seg.split("_")).capitalize()
         
         fig.add_trace(go.Scatter(
             x=data.index,
-            y=data['AoA'],
+            y=data['eta'],
             name=seg_name),
             row=1, col=1)
         
         fig.add_trace(go.Scatter(
             x=data.index,
-            y=data['CL'],
+            y=data['Lift'],
+            name=seg_name,
+            showlegend=False),
+            row=2, col=1)
+        
+        fig.add_trace(go.Scatter(
+            x=data.index,
+            y=data['Thrust'],
             name=seg_name,
             showlegend=False),
             row=1, col=2)
         
         fig.add_trace(go.Scatter(
             x=data.index,
-            y=data['CD'],
+            y=data['Drag'],
             name=seg_name,
             showlegend=False),
             row=2, col=2)
-        
-        fig.add_trace(go.Scatter(
-            x=data.index,
-            y=data['L_D'],
-            name=seg_name,
-            showlegend=False),
-            row=2, col=1)        
-    
-    fig.update_yaxes(title_text='AoA (deg)', row=1, col=1)
-    fig.update_yaxes(title_text='CL', row=1, col=2)
-    fig.update_yaxes(title_text='CD', row=2, col=2)
-    fig.update_yaxes(title_text='L/D', row=2, col=1)
+                                        
+    fig.update_yaxes(title_text='Throttle', row=1, col=1)
+    fig.update_yaxes(title_text='Lift (N)', row=2, col=1)
+    fig.update_yaxes(title_text='Thrust (N)', row=1, col=2)
+    fig.update_yaxes(title_text='Drag (N)', row=2, col=2)
     
     fig.update_xaxes(title_text='Time (min)', row=2, col=1)
-    fig.update_xaxes(title_text='Time (min)', row=2, col=2)    
+    fig.update_xaxes(title_text='Time (min)', row=2, col=2)        
 
     # Set overall figure layout style and legend title
     fig.update_layout(
         width=width, height=height,
         legend_title_text='Segment',
-        title_text = 'Aerodynamic Coefficients'
+        title_text = 'Aerodynamic Forces'
     )
 
     fig = plot_style(fig)
@@ -139,5 +140,6 @@ def plot_aerodynamic_coefficients(results,
 
     if save_figure:
         save_plot(fig, save_filename, file_type)
+
 
     return
