+++ trunk/SUAVE/Visualization/Performance/Aerodynamics/Vehicle/plot_stability_coefficients.py
@@ -1,5 +1,5 @@
 ## @ingroup Visualization-Performance-Aerodynamics
-# plot_aerodynamic_forces.py
+# plot_stability_coefficients.py
 # 
 # Created:    Nov 2022, E. Botero
 # Modified:   
@@ -18,24 +18,22 @@ import plotly.graph_objects as go
 from plotly.subplots import make_subplots
 
 # ---------------------------------------------------------------------- 
-#   Aerodynamic Forces
+#   Stability Coefficients
 # ---------------------------------------------------------------------- 
 
 ## @ingroup Visualization-Performance-Aerodynamics
-def plot_aerodynamic_forces(results,
-                            save_figure=False,
-                            save_filename="Aerodynamic_Forces",
-                            file_type=".png",
-                            width = 1600, height = 665,
-                            *args, **kwargs):
-    """This plots the aerodynamic forces
+def plot_stability_coefficients(results,
+                                save_figure=False,
+                                save_filename="Stability_Coefficents",
+                                file_type=".png",
+                                width = 1400, height =800,
+                                *args, **kwargs):
+    """This plots the static stability characteristics of an aircraft
     
     Assumptions:
     None
     
     Source:
-    
-    Deprecated SUAVE Mission Plots Functions
 
     Created:    Mar 2020, M. Clarke
     Modified:   Apr 2020, M. Clarke
@@ -43,87 +41,89 @@ def plot_aerodynamic_forces(results,
                 Apr 2021, M. Clarke
                 Dec 2021, S. Claridge
     
-    Inputs:
-    results.segments.condtions.frames
-         body.thrust_force_vector
-         wind.lift_force_vector
-         wind.drag_force_vector
-         
-    Outputs:
-    Plots
     
+    Inputs:
+    results.segments.conditions.stability.
+       static
+           CM
+           Cm_alpha
+           static_margin
+       aerodynamics.
+           angle_of_attack
+    Outputs:
+    
+    Plots
     Properties Used:
     N/A
     """
     
-    # Create empty dataframe to be populated by the segment data
-    plot_cols = ['Thrust',
-                 'Lift',
-                 'Drag',
-                 'eta',
+    plot_cols = ['CM',
+                 'CM_a',
+                 'SM',
+                 'AoA',
                  'Segment']
 
-    df = pd.DataFrame(columns=plot_cols)    
+    df = pd.DataFrame(columns=plot_cols)
+    
 
     for segment in results.segments.values():
-        time   = segment.conditions.frames.inertial.time[:,0] / Units.min
-        Thrust = segment.conditions.frames.body.thrust_force_vector[:,0]
-        Lift   = -segment.conditions.frames.wind.lift_force_vector[:,2]
-        Drag   = -segment.conditions.frames.wind.drag_force_vector[:,0]
-        eta    = segment.conditions.propulsion.throttle[:,0]
+        time     = segment.conditions.frames.inertial.time[:,0] / Units.min
+        cm       = segment.conditions.stability.static.CM[:,0]
+        cm_alpha = segment.conditions.stability.static.Cm_alpha[:,0]
+        SM       = segment.conditions.stability.static.static_margin[:,0]
+        aoa      = segment.conditions.aerodynamics.angle_of_attack[:,0] / Units.deg
         
-        # Assemble data into temporary holding data frame
         segment_frame = pd.DataFrame(
-            np.column_stack((Thrust,
-                             Lift,
-                             Drag,
-                             eta)),
-            columns = plot_cols[:-1], index=time
+        np.column_stack((cm,
+                         cm_alpha,
+                         SM,
+                         aoa)),
+        columns = plot_cols[:-1], index=time
         )
+
         segment_frame['Segment'] = [segment.tag for i in range(len(time))]
-
+    
         # Append to collecting data frame
-        df = df.append(segment_frame)        
-
+        df = df.append(segment_frame)             
 
     # Set plot parameters
     fig = make_subplots(rows=2, cols=2)
-    
+
     # Add traces to the figure for each value by segment
     for seg, data in df.groupby("Segment", sort=False):
-        seg_name = ' '.join(seg.split("_")).capitalize()
+        seg_name = ' '.join(seg.split("_")).capitalize()  
         
         fig.add_trace(go.Scatter(
             x=data.index,
-            y=data['eta'],
+            y=data['AoA'],
             name=seg_name),
-            row=1, col=1)
+            row=1, col=1)    
         
         fig.add_trace(go.Scatter(
             x=data.index,
-            y=data['Lift'],
+            y=data['CM'],
             name=seg_name,
             showlegend=False),
-            row=2, col=1)
-        
+            row=1, col=2)        
+
         fig.add_trace(go.Scatter(
             x=data.index,
-            y=data['Thrust'],
+            y=data['CM_a'],
             name=seg_name,
             showlegend=False),
-            row=1, col=2)
-        
+            row=2, col=1)        
+
         fig.add_trace(go.Scatter(
             x=data.index,
-            y=data['Drag'],
+            y=data['SM'],
             name=seg_name,
             showlegend=False),
-            row=2, col=2)
-                                        
-    fig.update_yaxes(title_text='Throttle', row=1, col=1)
-    fig.update_yaxes(title_text='Lift (N)', row=2, col=1)
-    fig.update_yaxes(title_text='Thrust (N)', row=1, col=2)
-    fig.update_yaxes(title_text='Drag (N)', row=2, col=2)
+            row=2, col=2)        
+
+    fig.update_yaxes(title_text='AoA (deg)', row=1, col=1)
+    fig.update_yaxes(title_text='$C_M$', row=1, col=2)
+    fig.update_yaxes(title_text='$C_M\alpha$', row=2, col=1)
+    fig.update_yaxes(title_text='Static Margin (%)', row=2, col=2)
     
     fig.update_xaxes(title_text='Time (min)', row=2, col=1)
     fig.update_xaxes(title_text='Time (min)', row=2, col=2)        
@@ -132,7 +132,7 @@ def plot_aerodynamic_forces(results,
     fig.update_layout(
         width=width, height=height,
         legend_title_text='Segment',
-        title_text = 'Aerodynamic Forces'
+        title_text = 'Stability Coefficients'
     )
 
     fig = plot_style(fig)
@@ -140,6 +140,5 @@ def plot_aerodynamic_forces(results,
 
     if save_figure:
         save_plot(fig, save_filename, file_type)
-
 
     return
