+++ trunk/SUAVE/Visualization/Performance/Aerodynamics/__init__.py
@@ -1,0 +1,7 @@
+## @defgroup Visualization-Performance-Aerodynamics Aerodynamics
+# Plotting Functions for Aerodynamics
+# @ingroup Visualization-Performance
+
+from . import Airfoil  
+from . import Rotor
+from . import Vehicle
