+++ trunk/SUAVE/Visualization/Performance/Energy/Common/plot_disc_power_loading.py
@@ -23,7 +23,7 @@ def plot_disc_power_loading(results,
                             save_figure=False,
                             save_filename="Disc_Power_Loading",
                             file_type = ".png",
-                            width = 1600, height = 800,
+                            width = 1400, height = 800,
                             *args, **kwargs):
     """Plots propeller disc and power loadings
