--- trunk/SUAVE/Visualization/Performance/Rotor_Aerodynamics/__init__.py
@@ -1,7 +1,0 @@
-## @defgroup Visualization-Performance-Rotor_Aerodynamics Performance 
-# @ingroup Visualization
- 
-
-from .plot_rotor_disc_inflow      import plot_rotor_disc_inflow 
-from .plot_rotor_disc_performance import plot_rotor_disc_performance
-from .plot_rotor_performance      import plot_rotor_performance
