--- trunk/SUAVE/Visualization/Performance/Vehicle_Aerodynamics/__init__.py
@@ -1,9 +1,0 @@
-## @defgroup Visualization-Performance-Vehicle_Aerodynamics Aerodynamics
-# Plotting Functions for Vehicle_Aerodynamics
-# @ingroup Visualization-Performance
-
-from .plot_aerodynamic_coefficients     import plot_aerodynamic_coefficients
-from .plot_aerodynamic_forces           import plot_aerodynamic_forces
-from .plot_drag_components              import plot_drag_components
-from .plot_lift_distribution            import plot_lift_distribution
-from .plot_stability_coefficients       import plot_stability_coefficients
