+++ trunk/SUAVE/Visualization/Performance/__init__.py
@@ -2,9 +2,7 @@
 # Visualization contains functions for generating common figures
 # @ingroup Visualization 
 
-from . import Vehicle_Aerodynamics
-from . import Rotor_Aerodynamics
-from . import Airfoil_Aerodynamics
+from . import Aerodynamics
 from . import Common
 from . import Energy
 from . import Mission
